Remove castOutputs from QuerySpec
This removes the casts from the analyzer and instead injects a
projection with casts in the planner.

Doing the casts in the planner has the advantage that we only need to
cast the outputs and don't have to adopt the group keys or order by
items.

Main motivation of this change is to to reduce the usage of the
`QuerySpec` to eventually get rid of it.